requirements.txt
langchain==0.1.9
langchain-openai==0.0.8
chromadb==0.4.22
python-dotenv==1.0.1
unstructured[md]==0.11.8
pdf2image==1.17.0
pytesseract==0.3.10
python-magic==0.4.27
tabulate==0.9.0 